requirements.txt
ffmpeg-python==0.2.0
keyboard==0.13.5
pytube==12.1.0